# apps/server/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/invoices": {
      "get": {
        "operationId": "listInvoices",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceListResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoices/{id}": {
      "get": {
        "operationId": "getInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/projects/{id}/invoices": {
      "post": {
        "operationId": "createProjectInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoices/{id}/actions/revise": {
      "post": {
        "operationId": "reviseInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoice-versions/{id}/actions/prepare": {
      "post": {
        "operationId": "prepareInvoiceVersion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoice-versions/{id}/actions/post": {
      "post": {
        "operationId": "postInvoiceVersion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
...
    "/api/v1/invoices/{id}/public-links": {
      "post": {
        "operationId": "createInvoicePublicLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoicePublicLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoicePublicLinkDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoices/{id}/public-links/{linkId}/actions/revoke": {
      "post": {
        "operationId": "revokeInvoicePublicLink",
        "parameters": [
          {
            "name": "linkId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoices/{id}/adjustments": {
      "post": {
        "operationId": "createInvoiceAdjustment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceAdjustmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoice-adjustments/{id}/actions/approve": {
      "post": {
        "operationId": "approveInvoiceAdjustment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoice-adjustments/{id}/actions/post": {
      "post": {
        "operationId": "postInvoiceAdjustment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoice-adjustments/{id}/actions/reverse": {
      "post": {
        "operationId": "reverseInvoiceAdjustment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCorrectionReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoices/{id}/actions/void": {
      "post": {
        "operationId": "voidInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCorrectionReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/invoices/{id}/actions/replace": {
      "post": {
        "operationId": "replaceInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
...
    "/api/v1/public/invoices/{token}": {
      "get": {
        "operationId": "getPublicInvoice",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicInvoiceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Public Invoices"]
      }
    },
...
  },
  "info": {
    "title": "Lincoln Dirt and Gravel API",
    "description": "Lincoln Dirt and Gravel operating system API",
    "version": "1.7.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "InvoiceListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceSummaryDto"
            }
          }
        },
        "required": ["items"]
      },
...
      "InvoiceDeliveryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "invoiceVersionId": {
            "type": "string",
            "format": "uuid"
          },
          "channel": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "destination": {
            "type": "object",
            "additionalProperties": true
          },
          "attemptedAt": {
            "type": "string",
            "format": "date-time"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "viewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "failureReason": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["id", "invoiceVersionId", "channel", "status", "destination", "attemptedAt"]
      },
...
      "InvoiceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "invoiceNumber": {
            "type": "string"
          },
          "invoiceType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "jobId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "projectNumber": {
            "type": "string"
          },
          "jobNumber": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string"
          },
          "totalCents": {
            "type": "number",
            "nullable": true
          },
          "outstandingBalanceCents": {
            "type": "number"
          },
          "replacesInvoiceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "replacedByInvoiceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "issueDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "adjustments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceAdjustmentDto"
            }
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceVersionDto"
            }
          },
          "deliveries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDeliveryDto"
            }
          },
          "publicLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoicePublicLinkSummaryDto"
            }
          },
          "voidedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "voidReason": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "invoiceNumber",
          "invoiceType",
          "status",
          "currency",
          "projectId",
          "projectNumber",
          "customerName",
          "outstandingBalanceCents",
          "updatedAt",
          "adjustments",
          "versions",
          "deliveries",
          "publicLinks"
        ]
      },
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "invoiceType": {
            "type": "string",
            "enum": ["deposit", "final", "additional_charge", "credit_memo"]
          },
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "dueInDays": {
            "type": "number",
            "default": 14,
            "maximum": 365,
            "minimum": 0
          }
        },
        "required": ["invoiceType"]
      },
...
      "CreateInvoicePublicLinkDto": {
        "type": "object",
        "properties": {
          "recipient": {
            "type": "string",
            "maxLength": 320
          },
          "expiresInDays": {
            "type": "number",
            "default": 14,
            "maximum": 30,
            "minimum": 1
          }
        },
        "required": ["recipient"]
      },
      "InvoicePublicLinkDto": {
        "type": "object",
        "properties": {
          "invoice": {
            "$ref": "#/components/schemas/InvoiceDto"
          },
          "linkId": {
            "type": "string",
            "format": "uuid"
          },
          "customerPath": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["invoice", "linkId", "customerPath", "token", "expiresAt"]
      },
      "CreateInvoiceAdjustmentDto": {
        "type": "object",
        "properties": {
          "adjustmentType": {
            "type": "string",
            "enum": [
              "additional_charge",
              "credit",
              "tax_adjustment",
              "write_off",
              "due_date_extension"
            ]
          },
          "direction": {
            "type": "string",
            "enum": ["debit", "credit"]
          },
          "amountCents": {
            "type": "number",
            "minimum": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 1000
          },
          "effectiveAt": {
            "type": "string",
            "format": "date-time"
          },
          "sourceType": {
            "type": "string",
            "maxLength": 50
          },
          "sourceId": {
            "type": "string",
            "format": "uuid"
          },
          "newDueDate": {
            "type": "string",
            "format": "date"
          }
        },
        "required": ["adjustmentType", "amountCents", "reason"]
      },
      "InvoiceCorrectionReasonDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": ["reason"]
      },
      "ReplaceInvoiceDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000
          },
          "dueInDays": {
            "type": "number",
            "maximum": 365,
            "minimum": 0
          }
        },
        "required": ["reason"]
      },
...
      "PublicInvoiceDto": {
        "type": "object",
        "properties": {
          "invoiceNumber": {
            "type": "string"
          },
          "invoiceType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "projectNumber": {
            "type": "string"
          },
          "jobNumber": {
            "type": "string",
            "nullable": true
          },
          "issueDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "versionNumber": {
            "type": "number"
          },
          "subtotalCents": {
            "type": "number"
          },
          "taxCents": {
            "type": "number"
          },
          "totalCents": {
            "type": "number"
          },
          "appliedCents": {
            "type": "number"
          },
          "outstandingBalanceCents": {
            "type": "number"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineItemDto"
            }
          },
          "adjustments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicInvoiceAdjustmentDto"
            }
          },
          "deliveries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicInvoiceDeliveryDto"
            }
          }
        },
        "required": [
          "invoiceNumber",
          "invoiceType",
          "status",
          "currency",
          "customerName",
          "projectNumber",
          "versionNumber",
          "subtotalCents",
          "taxCents",
          "totalCents",
          "appliedCents",
          "outstandingBalanceCents",
          "lines",
          "adjustments",
          "deliveries"
        ]
      },
...
    }
  }
}